+++ packages/element-library/src/media/util.js
@@ -17,7 +17,7 @@
 /**
  * External dependencies
  */
-import { css } from 'styled-components';
+import styled, { css } from 'styled-components';
 
 export const mediaWithScale = css`
   width: ${({ width }) => `${width}px`};
@@ -31,3 +31,34 @@ export function getMediaWithScaleCss({ width, height, offsetX, offsetY }) {
   // no other apparent way to execute interpolate `mediaWithScale` dynamically.
   return `width:${width}px; height:${height}px; left:${-offsetX}px; top:${-offsetY}px;`;
 }
+
+const videoWithScale = css`
+  width: ${({ width }) => `${width}px`};
+  left: ${({ offsetX }) => `${-offsetX}px`};
+  top: ${({ offsetY }) => `${-offsetY}px`};
+  max-width: ${({ isBackground }) => (isBackground ? 'initial' : null)};
+`;
+
+export const getBackgroundStyle = () => {
+  return {
+    minWidth: '100%',
+    minHeight: '100%',
+    maxWidth: 'initial',
+  };
+};
+
+export const Video = styled.video`
+  position: absolute;
+  max-width: initial;
+  max-height: initial;
+  height: ${({ height }) => `${height}px`};
+  background-image: ${({ poster }) => `url("${poster}")`};
+  ${videoWithScale}
+`;
+
+export const VideoImage = styled.img`
+  position: absolute;
+  max-height: initial;
+  object-fit: contain;
+  ${videoWithScale}
+`;
